Adding a block-rescue section around downloading videos
This commit should add new functionality to the playbook. There is
now a try-catch like method where it will try to download the full
video and if that fails, then it will try to get the thumbnail. The
thumbnail doesn't always match the picture the user wanted but I
wanted this playbook to just give something instead of failing
outright.